# src/main/java/io/umiverse/gateway/openapi/model/AskAcceptRequest.java
package io.umiverse.gateway.openapi.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import javax.validation.Valid;
import javax.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import javax.annotation.Generated;

/**
 * AskAcceptRequest
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-08-03T15:10:02.892036+08:00[Asia/Shanghai]")
public class AskAcceptRequest {

  @JsonProperty("nftId")
  private Integer nftId;

  public AskAcceptRequest nftId(Integer nftId) {
    this.nftId = nftId;
    return this;
  }

  /**
   * NFT 数据库对应ID
   * @return nftId
  */
  @NotNull 
  @Schema(name = "nftId", description = "NFT 数据库对应ID", required = true)
  public Integer getNftId() {
    return nftId;
  }

  public void setNftId(Integer nftId) {
    this.nftId = nftId;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    AskAcceptRequest askAcceptRequest = (AskAcceptRequest) o;
    return Objects.equals(this.nftId, askAcceptRequest.nftId);
  }

  @Override
  public int hashCode() {
    return Objects.hash(nftId);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class AskAcceptRequest {\n");
    sb.append("    nftId: ").append(toIndentedString(nftId)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}